update explain state

+++ src/modules/task-manager/dialogs/task-update-state-dialog.vue
@@ -13,12 +13,8 @@
         <v-container fluid px-5 py-2>
           <v-row>
             <v-col cols="12" class="pa-2">
-              <task-state-select :value.sync="state" label="Trạng thái" />
-              <date-picker-input
-                :value.sync="executedDate"
-                :rules="$appRules.taskExecutedDate"
-                label="Ngày thực hiện"
-              />
+              <task-state-select :includes="taskStateIncludes" :value.sync="state" label="Trạng thái" />
+              <date-picker-input :value.sync="startedDate" :rules="$appRules.taskStartedDate" label="Ngày thực hiện" />
               <app-text-field v-model="explain" :rules="$appRules.taskExplain" label="Diễn giải trạng thái" />
               <app-file-input hide-details label="File đính kèm" />
             </v-col>
@@ -39,7 +35,8 @@
 
 <script lang="ts">
 import { AppProvider } from '@/app-provider'
-import { createTaskBody, TaskModel, TaskStateType } from '@/models/task-model'
+import { RequestModel } from '@/models/request-model'
+import { createTaskBody, getLastRequest, TaskModel, TaskStateType } from '@/models/task-model'
 import { authStore } from '@/stores/auth-store'
 import { Component, Inject, Prop, PropSync, Ref, Vue, Watch } from 'vue-property-decorator'
 
@@ -54,16 +51,30 @@ export default class TaskUpdateStateDialog extends Vue {
   @PropSync('value', { type: Boolean, default: false }) syncedValue!: boolean
   @Ref('form') form: any
   @Prop() task: TaskModel
+  @Prop() isUpdateTask!: boolean
 
   code = ''
   state: TaskStateType = null
   explain = ''
-  executedDate: string = null
+  startedDate: string = null
+  taskStateIncludes: TaskStateType[] = ['todo', 'doing', 'waiting', 'done']
+  request: RequestModel = null
 
-  @Watch('task', { immediate: true }) onTaskChanged(val: TaskModel) {
+  @Watch('value', { immediate: true }) onValueChanged(val: string) {
     if (val) {
-      this.code = val.code
-      this.state = val.state
+      if (this.isUpdateTask) {
+        this.code = this.task.code
+        this.state = this.task.state
+      } else {
+        const lastRequest = getLastRequest(this.task)
+        if (!lastRequest) this.syncedValue = false
+        else {
+          this.request = lastRequest
+          this.state = lastRequest.type
+          this.startedDate = lastRequest.startedDate
+          this.explain = lastRequest.description
+        }
+      }
     }
   }
 
@@ -71,21 +82,31 @@ export default class TaskUpdateStateDialog extends Vue {
     if (this.form.validate()) {
       try {
         const api = this.providers.api
+        let request
+        if (this.request) {
+          request = await api.request.update(this.request.id, {
+            type: this.state,
+            startedDate: this.startedDate,
+            description: this.explain
+          })
+        } else {
+          request = await api.request.create({
+            // title, files, approver
+            description: this.explain,
+            type: this.state,
+            startedDate: this.startedDate,
+            requestor: authStore.comrade.id,
+            task: this.task.id
+          })
+        }
 
-        const request = await api.request.create({
-          // title, files, approver
-          description: this.explain,
-          type: this.state,
-          requestor: authStore.comrade.id,
-          task: this.task.id
-        })
         try {
           const modifyTask = await api.task.update(
             this.task.id,
             createTaskBody(this.task, {
               state: this.state,
               status: this.state === 'done' ? 'approving' : null,
-              data: { ...(this.task.data ?? {}), explain: this.explain }
+              explainState: this.explain
             })
           )
           this.$emit('success', modifyTask)
@@ -93,7 +114,7 @@ export default class TaskUpdateStateDialog extends Vue {
           this.form.reset()
           this.providers.snackbar.updateSuccess()
         } catch (error) {
-          await api.request.delete(request.id)
+          if (!this.request) await api.request.delete(request.id)
           throw error
         }
       } catch (error) {
